+++ aalekh-gradle/src/main/kotlin/com/aalekh/aalekh/gradle/AalekhSettingsPlugin.kt
@@ -114,6 +114,7 @@ public class AalekhSettingsPlugin : Plugin<Settings> {
                 task.ruleEntries.set(extension.rulesConfig.entries)
                 task.forbidEntries.set(extension.forbiddenEntries)
                 task.reachabilityEntries.set(extension.rulesConfig.reachabilityEntries)
+                task.sourceSetEntries.set(extension.rulesConfig.sourceSetEntries)
                 task.teamEntries.set(rootProject.provider {
                     extension.teamOwnership.toInputString()
                 })
@@ -145,6 +146,7 @@ public class AalekhSettingsPlugin : Plugin<Settings> {
                 task.ruleEntries.set(extension.rulesConfig.entries)
                 task.forbidEntries.set(extension.forbiddenEntries)
                 task.reachabilityEntries.set(extension.rulesConfig.reachabilityEntries)
+                task.sourceSetEntries.set(extension.rulesConfig.sourceSetEntries)
                 task.baselineFile.set(
                     extension.baselineFile.map { rootProject.layout.projectDirectory.file(it) }
                 )
@@ -189,6 +191,7 @@ public class AalekhSettingsPlugin : Plugin<Settings> {
                 task.ruleEntries.set(extension.rulesConfig.entries)
                 task.forbidEntries.set(extension.forbiddenEntries)
                 task.reachabilityEntries.set(extension.rulesConfig.reachabilityEntries)
+                task.sourceSetEntries.set(extension.rulesConfig.sourceSetEntries)
                 task.baselineOutputFile.set(
                     extension.baselineFile.map { rootProject.layout.projectDirectory.file(it) }
                 )